+++ buildSrc/src/main/groovy/com/uber/okbuck/composer/AndroidLibraryRuleComposer.groovy
@@ -5,7 +5,7 @@ import com.uber.okbuck.core.model.AndroidTarget
 import com.uber.okbuck.core.model.Target
 import com.uber.okbuck.generator.RetroLambdaGenerator
 import com.uber.okbuck.rule.AndroidLibraryRule
-import com.uber.okbuck.block.PostProcessClassessCommands
+import com.uber.okbuck.printable.PostProcessClassessCommands
 
 final class AndroidLibraryRuleComposer extends AndroidBuckRuleComposer {
