+++ buildSrc/src/main/groovy/com/uber/okbuck/composer/ExopackageAndroidLibraryRuleComposer.groovy
@@ -3,7 +3,7 @@ package com.uber.okbuck.composer
 import com.uber.okbuck.core.model.AndroidAppTarget
 import com.uber.okbuck.generator.RetroLambdaGenerator
 import com.uber.okbuck.rule.ExopackageAndroidLibraryRule
-import com.uber.okbuck.block.PostProcessClassessCommands
+import com.uber.okbuck.printable.PostProcessClassessCommands
 
 final class ExopackageAndroidLibraryRuleComposer extends AndroidBuckRuleComposer {
