+++ buildSrc/src/main/groovy/com/uber/okbuck/composer/JavaTestRuleComposer.groovy
@@ -3,7 +3,7 @@ package com.uber.okbuck.composer
 import com.uber.okbuck.core.model.JavaLibTarget
 import com.uber.okbuck.generator.RetroLambdaGenerator
 import com.uber.okbuck.rule.JavaTestRule
-import com.uber.okbuck.block.PostProcessClassessCommands
+import com.uber.okbuck.printable.PostProcessClassessCommands
 
 final class JavaTestRuleComposer extends JavaBuckRuleComposer {
