+++ buildSrc/src/main/groovy/com/uber/okbuck/rule/AndroidRule.groovy
@@ -1,6 +1,6 @@
 package com.uber.okbuck.rule
 
-import com.uber.okbuck.block.PostProcessClassessCommands
+import com.uber.okbuck.printable.PostProcessClassessCommands
 import org.apache.commons.lang.StringUtils
 
 abstract class AndroidRule extends BuckRule {
