+++ buildSrc/src/main/groovy/com/uber/okbuck/rule/ExopackageAndroidLibraryRule.groovy
@@ -1,6 +1,6 @@
 package com.uber.okbuck.rule
 
-import com.uber.okbuck.block.PostProcessClassessCommands
+import com.uber.okbuck.printable.PostProcessClassessCommands
 
 final class ExopackageAndroidLibraryRule extends BuckRule {
     private final String mAppClass
